Physical-AI-TextBook/rag-backend/requirements-vercel.txt
# FastAPI - minimal
fastapi>=0.109.0
uvicorn>=0.27.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0

# Vector DB
qdrant-client>=1.7.1

# Google Gemini AI
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
httpx>=0.26.0

# SSE for streaming
sse-starlette>=1.8.2
